scrapper/requirements.txt
selenium
webdriver-manager
kafka-python
python-dotenv
requests
